core: fix redundant u64 casts in crash handler for Windows clang-tidy
Windows clang-tidy (readability-redundant-casting) rejects the nested
(u64)(uintptr_t) cast because uintptr_t is unsigned long long (== u64)
there; cast the void* address directly, matching the stacktrace fix in
0b35808. Also drop the redundant (u64) wrapper on the Win32
ExceptionInformation path (ULONG_PTR is already u64 on x64).

+++ core/crash.c
@@ -137,7 +137,7 @@ static void crash_print_frames(void) {
 		crash_append_cstr("#");
 		crash_append_u64(i);
 		crash_append_cstr(" ");
-		crash_append_hex((u64)(uintptr_t)crash_frames[i].address);
+		crash_append_hex((u64)crash_frames[i].address);
 		crash_append_cstr("\n");
 		crash_flush_line();
 	}
@@ -480,7 +480,7 @@ static void crash_print_report(EXCEPTION_POINTERS* info) {
 	if ((code == 0xC0000005u || code == 0xC0000006u) && record->NumberParameters >= 2u) {
 		crash_reset_line();
 		crash_append_cstr("faulting address: ");
-		crash_append_hex((u64)record->ExceptionInformation[1]);
+		crash_append_hex(record->ExceptionInformation[1]);
 		crash_append_cstr("\n");
 		crash_flush_line();
 	}
